server/requirements.txt
psycopg2-binary
nltk
summarizer
transformers
torch
bert-extractive-summarizer
pymystem3
